finished ia-yolo

+++ ia_yolo_foggy/core/data_make.py
@@ -54,12 +54,17 @@ def parse_annotation(annotation):
         A = 0.5  
         # beta = 0.08  
         beta = 0.01 * i + 0.05
+        img_name = '../../datasets/lpval/' + image_name + '_' + ("%.2f"%beta) + '.' + image_name_index
+
+        if os.path.exists(img_name):
+            print(f'{img_name} already exists.')
+            return
+
         size = math.sqrt(max(row, col)) 
         center = (row // 2, col // 2)  
         foggy_image = AddHaz_loop(img_f, center, size, beta, A)
         img_f = np.clip(foggy_image*255, 0, 255)
         img_f = img_f.astype(np.uint8)
-        img_name = '../../datasets/lpval/' + image_name + '_' + ("%.2f"%beta) + '.' + image_name_index
         #img_name = '../datasets/VOC/val/' + image_name + '_' + ("%.2f"%beta) + '.' + image_name_index
         cv2.imwrite(img_name, img_f)
